LayoutTests/http/wpt/resource-timing/rt-nextHopProtocol.worker-expected.txt


PASS Must have PerformanceObserver and PerformanceResourceTiming in DedicatedWorkerGlobalScope
PASS nextHopProtocol is expected to be 'http1/1'
